Update leaveStatus
There were some scenarios where HR would process a change before returning the employee from Leave. Causing the HR service not to send a request for a Store Transfer and/or Position Change.

That portion has been removed from the script, new procedure will recognize there was a change in Power Automate Web. Account will continue to be removed from Leave but if the data in AD doesn't match with HR's system it'll send the info over to the PositionChange/StoreTransfer flow.